Move store to the right folder
